server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
psutil==5.9.8
pyyaml==6.0.1
websockets==12.0
